Refactor network checker for js and add initial setup for wasmJs, js.
Refactored the `NetworkChecker` on the js platform to use `MutableStateFlow` and `StateFlow` for managing network status updates and modified the network status retrieval logic.
The NetworkType `UNKNOWN` is added.
Added initial files for wasmJs and js.
Added the initial `index.html` for both js and wasmJs,
Added the favicon.
Added the initial `main.kt` for js.
Updated `build.gradle.kts` to include build instructions for wasmJs and js.